move entity actions to recycle bin module
