+++ src/packages/core/entity-action/common/index.ts
@@ -2,7 +2,3 @@ export * from './duplicate/duplicate.action.js';
 export * from './delete/delete.action.js';
 export * from './move/move.action.js';
 export * from './sort-children-of/sort-children-of.action.js';
-export * from './trash/trash.action.js';
-
-export { UmbRestoreFromRecycleBinEntityAction } from './restore-from-recycle-bin/restore-from-recycle-bin.action.js';
-export { UmbEmptyRecycleBinEntityAction } from './empty-recycle-bin/empty-recycle-bin.action.js';
