+++ src/packages/core/recycle-bin/entity-action/empty-recycle-bin/empty-recycle-bin.action.kind.ts
@@ -1,23 +1,21 @@
-import { UMB_ENTITY_ACTION_DEFAULT_KIND_MANIFEST } from '../../default/default.action.kind.js';
+import { UMB_ENTITY_ACTION_DEFAULT_KIND_MANIFEST } from '../../../entity-action/default/default.action.kind.js';
 import type { UmbBackofficeManifestKind } from '@umbraco-cms/backoffice/extension-registry';
 
 export const manifest: UmbBackofficeManifestKind = {
 	type: 'kind',
-	alias: 'Umb.Kind.EntityAction.Trash',
-	matchKind: 'trash',
+	alias: 'Umb.Kind.EntityAction.RecycleBin.Empty',
+	matchKind: 'emptyRecycleBin',
 	matchType: 'entityAction',
 	manifest: {
 		...UMB_ENTITY_ACTION_DEFAULT_KIND_MANIFEST.manifest,
 		type: 'entityAction',
-		kind: 'trash',
-		api: () => import('./trash.action.js'),
-		weight: 1050,
+		kind: 'emptyRecycleBin',
+		api: () => import('./empty-recycle-bin.action.js'),
+		weight: 100,
 		forEntityTypes: [],
 		meta: {
 			icon: 'icon-trash',
-			label: 'Trash...',
-			itemRepositoryAlias: '',
-			recycleBinRepositoryAlias: '',
+			label: 'Empty Recycle Bin...',
 		},
 	},
 };
