+++ src/packages/core/recycle-bin/entity-action/empty-recycle-bin/manifests.ts
