+++ src/packages/core/recycle-bin/entity-action/index.ts
@@ -1,0 +1,3 @@
+export { UmbTrashEntityAction } from './trash/trash.action.js';
+export { UmbRestoreFromRecycleBinEntityAction } from './restore-from-recycle-bin/restore-from-recycle-bin.action.js';
+export { UmbEmptyRecycleBinEntityAction } from './empty-recycle-bin/empty-recycle-bin.action.js';
