+++ src/packages/core/recycle-bin/entity-action/restore-from-recycle-bin/manifests.ts
