+++ src/packages/core/recycle-bin/entity-action/restore-from-recycle-bin/restore-from-recycle-bin.action.ts
@@ -1,4 +1,4 @@
-import { UmbEntityActionBase } from '../../entity-action-base.js';
+import { UmbEntityActionBase } from '../../../entity-action/entity-action-base.js';
 import type { MetaEntityActionRestoreFromRecycleBinKind } from '@umbraco-cms/backoffice/extension-registry';
 
 export class UmbRestoreFromRecycleBinEntityAction extends UmbEntityActionBase<MetaEntityActionRestoreFromRecycleBinKind> {
