+++ src/packages/core/recycle-bin/entity-action/trash/manifests.ts
