+++ src/packages/core/recycle-bin/index.ts
@@ -1,4 +1,9 @@
 export { UmbRecycleBinRepositoryBase } from './recycle-bin-repository-base.js';
+export {
+	UmbTrashEntityAction,
+	UmbRestoreFromRecycleBinEntityAction,
+	UmbEmptyRecycleBinEntityAction,
+} from './entity-action/index.js';
 
 export type { UmbRecycleBinDataSource } from './recycle-bin-data-source.interface.js';
 export type { UmbRecycleBinRepository } from './recycle-bin-repository.interface.js';
